Update where platform override comes from and better naming.
