+++ how-to/common/client/src/modules/endpoints/endpoint-shapes.ts
@@ -1,4 +1,3 @@
-
 interface BaseEndpointDefinition {
 	id: string;
 }
