+++ how-to/common/client/src/modules/endpoints/local-storage/platform-storage-shapes.ts
