+++ how-to/customize-workspace/public/manifest.fin.json
@@ -75,7 +75,7 @@
 		"appProvider": {
 			"endpointIds": ["apps-get", "common-apps-get"],
 			"cacheDurationInSeconds": 10,
-			"cacheDurationInMinutes":0,
+			"cacheDurationInMinutes": 0,
 			"appAssetTag": "appasset",
 			"manifestTypes": [
 				"view",
@@ -89,8 +89,7 @@
 			]
 		},
 		"endpointProvider": {
-			"modules": [
-			],
+			"modules": [],
 			"endpoints": [
 				{
 					"id": "apps-get",
